VGPR Refactor: Don't round up to occupancy boundary or reserve temps in advance
Previous code would pad VGPR usage up to the next occupancy boundary
and pre-allocate those VGPR as temps.  New code does not
"round-up" so the kernel uses only the VGPRs it needs.  This
provides a more accurate view of the vgpr usage of the kernel.

Tensile has a decent VGPR register allocator that can already
handle growing the temps if needed, and