Replace call_once with local static variables, and add a TypedContractionInputs instantiation for HPA
I just verified that if we keep the `TypedContractionInputs<Half, Half, Half, Half, float, float>` and this PR is used, then rocBLAS links okay, and the HPA tests pass, even with the new client.

I went through the trouble of making a workaround, namely, to use `TypedContractionInputs<Half, Half, Half, Half, Half, Half>` with a `Tensile::ContractionProblem` of types `Half` and `Half` for input and output. I converted the `float` `alpha` and `beta` values to `Half` before storing them in `TypedContractionInputs`. The result is a SegFault.

There does not seem to be any way of specifying the compute type explicitly -- I wonder if it's using heuristics. I guess the only way it tells HGEMM from HPA is the types of alpha and beta.

If we do not include this change, then the current rocBLAS will get a linker error when building for the new client.

This should not affect any code which is not using the new client, since all this does (besides call_once->static) is add an extra instantiation of `TypedContractionInputs`.