[activation] Add ActivationHPA in YAML
1. ActivationHPA
If ActivationHPA is sett to true, the data type the activation is
using will be the same as high precision accumulate.

BFloat16 will automatically set ActivationHPA to True.
Single and double will automatically set ActivationHPA to False.

2. Naming
Now the name is properly generated from __str__ in class
ProblemType instead of padding "_ACTIVATION" at the end of the
kernel name.