Allow gfx900:xnack-
The miscommunication relating to whether gfx900 should be used with
or without gfx900:xnack- was never properly resolved and so many of the
libraries that depend on rocBLAS use gfx900:xnack- in their default
configuration.

This change allows both gfx900 and gfx900:xnack- to be accepted by
Tensile, thereby allowing users that are building from source to
choose whatever they prefer.