Improve finding the install location of ROCm
On Fedora, ROCm is part of the disto and is installed not to /opt/rocm
but rather to the system locations with /usr being the closest equivelent
to /opt/rocm.  Instead of requiring the user to set the enviroment variable
ROCM_PATH=/usr, check if the /opt/rocm dir is found and if it is not, set
to /usr based on detecting that we are running on Fedora.

This resolves a testing issue

___________ ERROR collecting Tensile/Tests/yaml_only/test_config.py ____________
/usr/lib/python3.13/site-packages/_pytest/runner.py:341: in from_call

...

E   FileNotFoundError: [Errno 2] No such file or directory: '/opt/rocm/bin/rocm_agent_enumerator'
=========================== short test summary info ============================
ERROR Tensile/Tests/yaml_only/test_config.py - FileNotFoundError: [Errno 2] N...

Signed-off-by: Tom Rix <trix@redhat.com>