[rocm-libraries] ROCm/rocm-libraries#1477 (commit 52eb6c1)
[Tensile] Fix gfx1032 missing in SupportedISA

## Motivation

gfx1032 is declared as supported in CMake and header files, but it is
missing in Tensile Python AsmCaps and SupportedISA. This causes error
when running TensileCreateLibrary with --verify-manifest
--lazy-library-loading.

Link: https://github.com/ROCm/TheRock/issues/1260

## Technical Details

The master file generation logic ignores any ISA that is not listed in
SupportedISA, but the manifest emits the entries anyways. This causes
problem with --verify-manifest when checking if all files listed in
manifest are correctly generated.

This issue only occurs with --lazy-library-loading, as without this
flag, Tensile wouldn't generate master logic files.

Rest of the logic files are generated correctly for gfx1032.

## Test Plan

Test generating gfx1032 Tensile library.

```
TensileCreateLibrary
--merge-files
--separate-architectures
--lazy-library-loading
--no-short-file-names
--verbose=0
--code-object-version=default
--library-format=msgpack
--architecture=gfx1030_gfx1032
--no-enumerate rocblas/library/src/blas3/Tensile/Logic/asm_full build/Tensile HIP --verify-manifest"
```

## Test Result

Command completes successfully with all library files emitted correctly.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.